Fail cf-push if secret variables failed to decrypt
If manifest template is missing variables jinja-cli will output an
exception to stderr and exit without printing anything to stdout.

This makes `cf push` run with an empty manifest, which it treats as
a valid call so it proceeds to deploy an app without any route or
environment variable configuration.

To avoid this we redirect stderr to sdout when generating a manifest,
so if an exception happens it is written to the manifest and since it's
invalid YAML `cf push` will fail the release right away.

The problem with this approach is that the exception no longer appears
in the command output or CI logs, so it's not obvious what is wrong with
the manifest generation.

An alternative approach would be to not specify application name in the
`cf push` command, making it invalid to use with an empty manifest. This
works for regular `cf push` since it will deploy all apps listed in the
manifest. However it doesn't fit with our zero-downtime release task,
since it requires apps to be pushed individually and has to name each
deployed app in advance.