Always return the download link containing the public subdomain
`url_for` generates a link based on the request Host header. This
means that if the app is accessed using the cloudapps.digital PaaS
subdomain it will return a download link based on that subdomain.

Since we want to always return links with the public download.*
domain, to avoid any potential errors we could either:

1. Only allow upload access to the app through the download.* domain,
   making all requests (not just download requests go through
   CloudFront)

2. Always use the download.* domain in the generated link, overriding
   request Host value.

While first approach would be easier to implement and doesn't require
a helper function to be used everywhere the link is returned it would
prevent us from moving the upload API from the public domain (eg to
reduce any latency added by CloudFront or to move it to a separate
PaaS app).