Scan document for viruses before uploading to S3
Upload endpoint will send the file to the notify-antivirus API and
check the result before uploading the file to S3.

If the file fails the virus scan an error is returned and the
upload is aborted.

Antivirus client uses a similar pattern to other requests wrappers.
Antivirus API doesn't return any specific errors at the moment, but
any messages in the 'error' JSON response key will get picked up by
the client when they are added.