Rename store/antivirus argument to document_stream
Since we need to reset the stream position AntivirusClient is calling
`.seek` at the end of the scan method, which means it doesn't accept
strings.

New argument name should clarify that both DocumentStore and
AntivirusClient expect a file-like object. This could be either
an actual file object, a StreamIO or an instance of Werkzeug's
FileStorage (which proxies `.seek` to `.stream.seek`).