Check uploaded document matches one of the allowed file types
Identifies the uploaded file type using libmagic and checks if it's
included in a list of allowed file types.

This allows us to set the correct mimetype on the S3 object and
stops users from uploading unsupported document types.

File identification is based on 'magic patterns' used by `file`
command, so while it should identify all valid PDF files as PDF
it doesn't guarantee that a file is valid and can be opened by
a PDF reader. This should be enough to inform users about our
supported document types and prevent accidental uploads of
unsupported formats, but doesn't protect against a user trying to
send an unsupported file by masking it as one of the allowed types.