Add some tolerance to the max document size check
MAX_CONTENT_LENGTH is checking the value of the request's Content-
Length header. This is generally slightly larger than the actual
uploaded file size since the multipart POST request body will include
data boundaries and Content-Type/Content-Disposition headers for
each file. It doesn't have a set overhead size since it depends on
the client implementation and the uploaded file name.

We don't really care if a file is slightly larger than 2MB, but we
need to allow any files that are below 2MB (since that's what users
would expect based on our documentation) so we add a 1KB buffer to
the content length check, which should cover the size of the
multipart form wrapper.