deploy with zero downtime deploy
keeps the same box, which means same service bindings etc.
commands don't work in makefiles so we've had to move it to the
procfile.

the rollback command previously deleted the new failed app and renamed
the old working app back to `app-name`, so that the next deploy would
be able to repeat the process. however, as the app name doesn't change
now this isn't needed and the command is just a no-op.