make testing for docx files more robust
some versions of `libmagic` report that docx files are actually of type
`application/zip` rather than `application/vnd.openxmlformats....`.
Which is technically true, they are zip archives. It appears that
libmagic on paas isn't as up to date as that on my mac, so this commit
adds code to do some manual checking.

If it reports as a zip archive, then we try and check the zip headers to
get the list of files (note, this doesn't actually uncompress anything).
If there's a `word/document.xml` file, then that's good enough for us to
assume it's a docx. This might lead to some weird false positives - the
next step would involve extracting `[Content_Types].xml` and parsing the
xml contained within - that's equivalent to what the libmagic library
does currently - see https://github.com/threatstack/libmagic/blob/master/magic/Magdir/msooxml