Don't let pyup update requirements.txt apart from for security
As per https://pyup.io/docs/bot/config/#specify-files, this means pyup
will not update requirements.txt under general circumstances. The
exception will be for security updates - although we won't merge the
security update as is, it will hopefully be the prompt for us to run
make freeze-requirements which will likely solve the insecure
requirement.