Force download of rtf files rather than inline
We have complaints from users that when they click the link to download
an rtf file it is shown in their browser as there is no
'Content-Disposition' header. This commit does similar to how we treat
CSVs and adds the header such that they will download for a user.

Note when testing. This problem does not appear to exist on my mac
(regardless of if I use Safari, Chrome or Edge) so it likely a
combination of browser and OS system so I have not been able to
replicate myself but have taken the same approach as we did for CSVs and
tested that it does at least not break things.