Allow uploading simple Excel spreadsheets
This has been requested by our users. While there are many Excel
formats we could support - macro-enabled, templated, binary - we
want to limit the set to those with an established a user need.

I've tested this manually with a sample .xlsx file provided by the
user, and can confirm it uploads/downloads as expected.