Revert "Allow uploading simple Excel spreadsheets"
This reverts commit a1b425e4187d0ee9a9787bf18c325b9810b5020d. After
testing with different browsers we discovered an issue where the file
extension is missing for .xlsx and .docx attachments downloaded via
MS Edge, which prevents them being opened [1]. Since .xlsx is a recent
addition, we decided to remove support for it until we fix the issue.

I've left the test for the MIME type in as it's still valid for us to
check we can infer it correctly, whether it's allowed or not.

[1]: https://github.com/alphagov/document-download-api/pull/162#issuecomment-803006654