Fix downloaded files not having an extension
Previously several files (docx, doc and odt) would download without
an extension in the filename in non-Chrome browsers, which makes it
hard to open them on Windows and other OSs. This switches all files
to be downloaded by default, except for formats which have reliable
in-browser preview: pdf and txt.

Unfortunately, there doesn't seem to be a way to specify a filename
without also forcing the file to download, since this is sent with
the Content-Disposition header that forces it. Ideally we would let
in-browser preview work for other formats (e.g. office docs) if the
user's browser has support for it. However, it doesn't seem like a
big issue to have the file downloaded.