Allow optional extension when downloading files
This fixes an issue with not being able to open some attachments
when downloaded with certain browsers on Windows.

Once this is in use we can check if traffic to the old route has
stopped, and then look to remove it.

Note this doesn't change the existing behaviour for CSV and RTF,
where we prevent in-browser preview entirely. For other file types,
modifying the route to include the extension means we still support
in-browser preview, if the browser does as well.