Change existing download URL to use extension
In the next commit we'll simplify deriving the extension from the
MIME type, which should avoid most of the boilerplate here.

We expect the extension to always be != None, since all ContentTypes
must be in the allowed list in the config. Inverting the dict in the
next commit will mean we raise a KeyError if this assumption becomes
false in future, making the issue easier to debug.