Include download URL in /check endpoint response
This will be used by the frontend instead of the current, coupled
string builder [1]. Using the new URL will automatically mean that
future downloads include the extension. Note that we intend for the
frontend "download" endpoint to start making an API call to /check.

Once the frontend is using the new data, we could also remove the
'file_exists' flag, since this is implied by 'document != None'. A
more conventional approach would just be to return a 404.

Note the format of the new response data is based on the "upload"
endpoint [2], and we can always add more data to it in future.

[1]: https://github.com/alphagov/document-download-frontend/blob/42f6c80b7df64a71e832d4c9f1e2a0c2df977e19/app/main/views/index.py#L54-L59
[2]: https://github.com/alphagov/document-download-api/blob/9d41451ad9c6b15ba5c93705e686ea60dd97c3ec/app/upload/views.py#L61