Bump utils to 55.1.4
This required bumping the minimum version of boto3 with:

    pip-compile -P awscli requirements.in

I haven't looked into the awscli/boto3/botocore changes due to the
high churn on those libraries. Given they're minor changes we can
assume they are benign. s3transfer changes are also benign [^1]

This is also a good opportunity to remove the redundant egg spec in
the requirements, like we have in other repos.

[^1]: https://github.com/boto/s3transfer/blob/develop/CHANGELOG.rst