Add HTTP `Referrer-Policy' as `no-referrer` to download views
The referrer-policy header defines when and what should be sent
in the `Referer` header when a http request is made from clicking
a link on a page.

https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Referrer-Policy

When serving document files in the browser we
do not currently set the `Referrer-Policy` header. This means the
browser default is generally followed. Over the past few years browser
defaults have been changed to `strict-origin-when-cross-origin`
however this is quite a recent change and before that most used
`no-referrer-when-downgrade`.
https://developer.chrome.com/blog/referrer-policy-new-chrome-default/

This means that when clicking on any links in pdf documents served
on documents.service.gov.uk, for users using older versions of browsers
we will send the full URL including query parameters as a `Referer`
header to these links. In practice this means we are
ending up with full urls for document download files in the GOV.UK
logs under the `Referer` header. This is bad.

To stop this happening, this commit sets `Referrer-Policy' as
`no-referrer` to stop this happening.

The main route that is affected is the `download_document` route.
However I have also set the header on the `get_document_metadata`
route to be over precautious. This is because the `get_document_metadata`
route is accessible to the public with the key and although it does
not return a link (as in no `<a>` tags) it can still return a URL.
The URL is unlikely to be clickable but let us take no risks.