Take base64 encoding into account in MAX_CONTENT_LENGTH check
The MAX_CONTENT_LENGTH was set to 2Mb with an extra 1Kb allowed to
account for the headers for the file. Base64 encoding increases the file
size by at least 133%* of the size of its source
(https://developer.mozilla.org/en-US/docs/Glossary/Base64#encoded_size_increase).
This meant that someone could use one of our clients to upload a file
of 1.6Mb (for example) and would get get an error - the file would pass the file
size check in the client since it's smaller than 2Mb. document-download-api
would then check the size of the base64 encoded file and give an error since
the file is now over the limit of 2Mb.

This change increases the file size check to take into account the fact
base64 encoding makes files larger.

* Note - file sizes can actually be increased a little more than 33% - they
can increase by 37% if line breaks are automatically inserted when
converting to base64 (RFC2045 requires a line break to be added after every
76 characters - https://datatracker.ietf.org/doc/html/rfc2045). I haven't
accounted for the extra 4% since none of our clients add line breaks. It's
possible that someone could use the REST API with line breaks added but
they would be very unlucky for the line breaks to push the file size
over the limit.