check octet-stream as well as zip in case they are docx/xlsx
paas python buildpack has an old version of libmagic (5.32, you can
check by running `file --version`).

We've seen a case where someone is trying to send a spreadsheet and
libmagic identifies it as 'Microsoft OOXML', but when you ask for its
mimetype it returns "application/octet-stream" (ie: a generic unknown
binary blob). Using more up-to-date versions of libmagic on my local
machine, it's identified as 'Microsoft Excel 2007+', with the mimetype
correctly being 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'.

So, to mitigate this, just check octet-streams as well as
application/zip to see if the might be a secret xlsx/docx file. As a
reminder, xlsx/docx are zip files under the hood and we check by
getting the file list for them and looking for a specific file that is
only found in xlsx/docx.