Upload hashed recipient email to S3 with document as metadata
If verification_email present in request json.

This is done so we can protect the documents from being inadvertently
shared to pages like internet archive or wayback machine. User will
have to confirm that they are the one meant to access the file
by entering their email address. This is not as strong as password
protection, but enough for the purpose.

For hashing algorithm, we chose Argon2ID based on:
https://cheatsheetseries.owasp.org/cheatsheets/Password_Storage_Cheat_Sheet.html#password-hashing-algorithms

The above docs also provided us with a choice of parameters to use.
Our memory usage is higher than our CPU usage so we chose the
parameters based on that. We also explicitly set hash length,
based on argon2 docs: https://argon2-cffi.readthedocs.io/en/stable/parameters.html