Add `/authenticate` endpoint for documents
This endpoint will be used to check that a user-submitted email address
matches the hashed one we have stored against the document in S3. If it
matches, we will return a signed blob of data encoding the service id
and document id which the frontend will store as a cookie on the user's
client. This will then be sent by the client when they try to access the
final file download URL, proving that they have successfully gone
through the authentication flow.