Add file extension to document metadata route
This is so the frontend can show what type of file the user will
be downloading.

I did consider sending the mimetype or the pretty file type from
the API instead of the file extension. I decided to go for the file
extension as
- the file extension is already information the API exposes when
providing a direct link to the file
- the pretty file type, ie 'text file' or 'Microsoft Word document'
feels too presentational to be the responsibility of the API
- the mimetype is more complex than the file extension to read and
understand and so this makes it simpler for the frontend to use