only allow access to document if signed data passed through in get
Only check this if the document has a hashed email in the metadata so
that old documents don't need to go through the flow. When the user
enters their email address at the prompt we put a signed blob for that
service_id/document_id in a cookie. When the user clicks the link, we'll
get the contents of that cookie and compare to the ID of the file being
accessed. If the ID matches (and the signature is valid and within 30
days) then we return the file.

Otherwise, we redirect the user back to the landing page so that they
can enter their email into the prompt again.