Return `available_until` from doc metadata endpoint
Now that the DocumentStore returns information about the last full day a
file is available until, let's also return that from the `/check` view
so that it's accessible to the frontend.

Note: document-download-api responses are cached in CloudFront for 24
hours. When deploying this, make sure to wait at least 24 hours before
deploying the related document-download-frontend change (or clear the
cache - but we don't have a good way to do this currently).

Co-authored-by: Pea Tyczyńska <pea.tyczynska@digital.cabinet-office.gov.uk>