Return 400 (not 500) on non-ascii document
At the moment, if a user sends us `file` data that isn't base64 encoded
(eg isn't in the ascii charset) then we are throwing a 500. Lately we've
had this error coming through in batches, which we assume is happening
because the user can't recover from the error because we aren't
explaining what's wrong. Let's return a better error telling them to
base64 encode the file.