Add dockerfile and ability to run locally using docker
The dockerfile copies https://github.com/alphagov/notifications-api/blob/main/docker/Dockerfile
very closely. The main differences are
- libmagic-dev is needed as mentioned in the README for this repo
- no entrypoint script like the API because there will only be one
  app that we run from this docker image, not multiple like the API

I've also copied the pattern of the API with scripts/run_locally_with_docker.sh
which seems OK but we can adapt later if we need to.

Worth pointing out that we don't have any instructions in the README
for this repo about what environment variables are needed to run
the app, in particular it appears to need at least `NOTIFY_ENVIRONMENT`
but I haven't tried to document this as part of this PR. For the
time being you will need to remember to pass these environment
variables into the make commands in the same way you need to
export them if trying to run flask without docker.

You can test this commit by
- make bootstrap-with-docker
- NOTIFY_ENVIRONMENT=development FLASK_APP=application.py FLASK_DEBUG=1 make run-flask-with-docker

In production, we will instead be running a different command in
the image, one using gunicorn more specifically.