Add support for blocking documents via tagging
At the moment we have a convoluted process for blocking access to
published documents that involves an S3 bucket/object policy. This isn't
efficient or scalable (bucket policies can be at most 20kb, so
eventually this will just plain be impossible).

Let's instead allow a tag to restrict access to the file. We check for
this tag before attempting any other operations on the S3 object.