Add new environment variable for `FRONTEND_HOSTNAME_INTERNAL`
When the apps are deployed in ECS we want to use the internal routes
where possible. We use the `FRONTEND_HOSTNAME` in two places, but only
one of these needs to use the public route. This adds a new variable,
`FRONTEND_HOSTNAME_INTERNAL`, which will be used when calling
document-download-frontend from within the app. When deployed on the
PaaS both environment variables should have the same value.