Override some pydantic errors to stay compatible
Pydantic does (some) of the input validation for us, eg around the
correct types and required values.

It generates errors for these automatically, but to ease the transition
and not change the API interface, let's override the errors we expect
and replace them with our "human-friendly" error messages.

I think we probably would want to drop these in the long-term, but that
will require updating the API callers first to be able to handle the new
responses.