Make tests consistently use http as the scheme
The two functions in `urls.py` were each returning the URL with a
different scheme in the tests. `get_frontend_download_url` returned
HTTPS URLs since it inherits the value of `HTTP_SCHEME` from the Config
base class. `get_direct_file_url` returned HTTP URLs - it uses `url_for`
which looks for `PREFERRED_URL_SCHEME` when not in a request context.
Since we don't set that, it uses the default of `http`.

This sets `HTTP_SCHEME` to `http` for the test config. This matches
the behaviour of other environments, where both functions return URLs
with the same scheme and also makes the changes in a later commit
easier to see.