Change how `get_direct_file_url` generates URLs
When the app is deployed on the PaaS, document-download-frontend calls
document-download-api on its public route and the `get_direct_file_url`
function generates a URL using that public route. On ECS,
document-download-frontend will use the private route, which results in
the URLs generated using the private domains which we don't want.

We now always use the api hostname from the `DOCUMENT_DOWNLOAD_API_HOSTNAME`
environment variable instead of letting `url_for` decide what to use.