Simplify hostnames for test config
This removes `SERVER_NAME` since the value was never used. It needed to
be defined so that we could use `url_for` outside of a request context.
Instead, we copy the approach in notifications-api and just use a test
context in the tests so that we can remove it.

It also changes `DOCUMENT_DOWNLOAD_API_HOSTNAME` to make it a little
clearer that it is just the `FRONTEND_HOSTNAME` prefixed with `download.`