Add support for JSON explicitly
On stack cflinuxfs3, libmagic doesn't detect JSON explicitly - it just
comes back as text/plain. Which means that we currently support sending
JSON files. But we tried to move to cflinuxfs4, which has a newer
version of libmagic that detects JSON as application/json, which we
haven't whitelisted.

Given it's just a text file, let's explicitly support this format.