Remove PaaS redis code
The REDIS_URL is now always passed in as an environment variable
ourselves, and we will no longer look at vcap services for it.

The "/1" behaviour we retain and move into the config.py file.

We also unbind from redis in the manifest, but will need to follow
this up with manually unbinding using the CLI as the manifest
won't unbind for us.