Change Dockerfile to use multi-stage builds
We want to be able to run the tests in Docker since we saw differences
in behaviour when running the app in Docker (on ECS) vs on the PaaS. The
Dockerfile cannot be used for this without changes since it doesn't
install any test dependencies.

This changes the Dockerfile to use multi-stage builds and to be as
consistent as possible with those we have in other apps.