Add tests for a longer CSV file
The test CSV file is very short (just one word) and we've seen slightly
different behaviour with longer CSV files. This adds a second, longer,
CSV file to the tests.

The behaviour of the two files in `test_document_upload_csv_handling` in
inconsistent but this matches the current behaviour.