Upgrade Dockerfile to use bookworm
When this was using bullseye we were using libmagic-dev version 5.39
which has a known bug where CSV files can be identified as "application/csv"
instead of registered "text/csv". By upgrading to Bookworm we get a
later version of libmagic-dev without the bug.

After this change, the tests pass with Docker.