Allow passing filename for documents
At the moment, when files are downloaded they simply have a UUID and
file extension. This isn't really useful for users as it won't ever help
them recognise what the file is.

Let's allow services to pass filenames through, so that when the file is
downloaded it has a service-customisable filename.

We will now send a 'Content-Disposition' header for all requests to the
'download document' endpoint. We force it as an attachment (which
triggers download in most browsers) for the same set of file extensions
as before - csv and rtf. For all others, we mark it as 'inline', telling
the browser to open it in a tab. But by specifying the content
disposition, we can include a filename that will auto-populate file
dialogs with the correct default filename.

If `filename` is passed in the data for uploading a new document, this
overrides any `is_csv` parameter. We may mark `is_csv` as deprecated
and prefer the `filename` data option - but the documentation needs to
be clear on this behaviour either way.