Limit custom filename length
AWS S3 (User defined) metadata is limited to 2 KB in PUT requests, so we
should make sure we're not using it all. I also want to not let
filenames use the majority of it so that there's still space for us to
use it for other things in the future, if we want.

By limiting filename length to 100 characters, we will in theory be
allowing up to 580 bytes for the filename, assuming the extension
(`.csv` takes up four and is always ASCII). The rest of the name could
be unicode characters, which through encoding are converted to \Uxxxx -
ie 6 bytes. So in the worst case, 96 UTF-8 characters will become 576
bytes. This is the extreme outlier, as I expect most filenames will be
fully or mainly ASCII, so normally up to 100 bytes.