expect REDIS_URL to contain a /1 domain
previously, we expected to be passed in a REDIS_URL that looked like
`rediss://some-domain:6379`, which we'd add `/1` to to use the second
database contained on the redis server.

This is a bit brittle, so switch to a new variable to define the entire
URL, which will always contain a /1 domain.

In terraform for our ECS instances, this is stored in
"redis_document_download_connection_string"

In credentials for paas, it's stored in REDIS_DOCUMENT_DOWNLOAD_URL.

In both environments, we pull it in to the existing REDIS_URL
environment variable