application: use EventletTimeoutMiddleware if using eventlet
also catch EventletTimeout exceptions, turning these into 504s,
which are the closest thing to an application timeout standard
codes seem to cover

also control the timeout time with a
`HTTP_SERVE_TIMEOUT_SECONDS` config/env var that we can control
from e.g. terraform, unifying this with the value we use for
gunicorn's `timeout` (which has little effect on the eventlet
`worker_class`, but controls a similar behaviour on non-async
workers), with the intention we start to respect this across all
our apps.

default to 30s timeout as this is what we've decided to set our
cloudfront/ALB timeouts to (we'll make this change shortly)

in EventletTimeoutMiddleware's current implementation, the
timeout gets cancelled as soon as the initial status response
is returned, so streaming responses shouldn't be affected any
more than normal responses

p.s. where are the rest of our error handlers for this app?