application: use EventletTimeoutMiddleware if using eventlet
also catch EventletTimeout exceptions, turning these into 504s,
which are the closest thing to an application timeout standard
codes seem to cover

30s timeout as this is what we've decided to set our
cloudfront/ALB timeouts to (we'll make this change shortly)

in EventletTimeoutMiddleware's current implementation, the
timeout gets cancelled as soon as the initial status response
is returned, so streaming responses shouldn't be affected any
more than normal responses

p.s. where are the rest of our error handlers for this app?