Add new allowed file formats - JPEG and PNG
For JPEG, I had to choose between .jpg and .jpeg file extension,
as they share the same mimetype, and we use mimetypes as
keys in our ALLOWED_FILE_TYPES dict. I chose .jpeg, since
.jpg is a legacy extension stemming from DOS system's restriction
on number of characters in an extension 30 years ago - so with
time it should be used less and less.