Allow users to upload both .jpeg and .jpg extensions
We already had two mimetypes for a single extension - ".rtf"
can have either "text/rtf" or "application/rtf" as its mimetype.

Now, in order to allow one mimetype with two possible extensions
- "image/jpeg" corresponding to either ".jpg" or "jpeg", we
had to change our data structure a bit - as neither extensions
nor mimetypes are unique anymore.

So we moved to a tuple of tuples, from which we create two dicts:
FILE_EXTENSIONS_TO_MIMETYPES, and MIME_TYPES_TO_FILE_EXTENSIONS.

When we convert this tuple of tuples to one of these dicts,
and the program runs into identical keys, it will choose the last pair
from the top.

So right now it will choose "application/rtf" for "rtf" extension,
and "jpeg" for "image/jpeg" mimetype.

Changing the order of the tuples will change the resulting dicts,
and helpfully it will also break unit tests.

Why it was important to allow both extensions?
Accepting both feels like a good example of the robustness principle
(“be conservative in what you send, be liberal in what you accept”)