download views: rework error handling and status codes
don't squash all store errors to 400 status code - if the intention
was to prevent information leak, it failed to do this because it
included the error string in the response. do this by making
DocumentStoreError a superclass of other store exception types so
exception handlers can continue to listen for the general
DocumentStoreError.

return 410 for keys that have a delete marker, 404 for documents
with no such key, 403 for documents that are blocked.

notably also catch NoSuchKey from check_for_blocked_document's
get_object_tagging calls and turn this into a DocumentNotFound
error, because in practise this will be the first problem a
request will encounter when trying to access a missing document.