DocumentStore: partially enforce expired document inaccessibility via app logic and Expiration metadata
this check was previously done in document-download-frontend, but
it really belongs in the document-download-api if anywhere, if
only because the api's download endpoints are publicly accessible
and a document denied access by just the frontend could always
just be downloaded via the api.

note this check still relies on the s3 bucket's lifecycle policies
being in place and implicitly the s3 object's last modified
timestamp. this will only "rescue" us in cases where a delete
marker has not yet been added by the lifecycle rile for some reason.