DocumentStore: add created-at tag when uploading documents
to an extent this is a "backup" of the creation-time information
we usually refer to s3's native timestamp for in case some
maintenance-related reason causes us to lose it, but we will
probably add logic to the app further down the line to use
the data from this tag for additional access enforcement.

timestamp will probably not match s3's native timestamp all
the time.