Ensure file checks are always called with a hash of the contents
This will allow us to cache this function further down the line.

Adds a second hash for integrity checking – makes sure the function is
never called with the wrong hash.

Hashing should be ridiculously fast for 2MB files so we don’t need to
worry about the performance implications of doing it once or twice.