Make `virus_free` a method
Before, virus free could theoretically:
- return `True`
- return `False`
- raise

In practice the way it was structured meant it could never return False.

This means calling code had to account for the `raise`/`False`
possibilities, even though the latter would never happen.

Changing it to a method makes it clearer that it’s doing work. It also
means it less unexpected when we change it to either:
- return `None` (success)
- raise (failed)