Use file extension to determine cache key
Only the file extension can change what MIME type is detected for the
file. The rest of the filename doesn’t matter.

This probably won’t result in a speedup (the dict lookup from extention
to MIME type is cheap) but it could result in slightly fewer cache keys
being used.